Move navbar to separate component, client
